% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_tbl}
\alias{as_tbl}
\title{Convert ig output to tibble}
\usage{
as_tbl(x)
}
\arguments{
\item{x}{Output from an ig function}
}
\value{
Print out of summary info and a tibble of the data.
}
\description{
Convert ig response objects into informative tibbles
}
