\name{ig-package}
\alias{ig-package}
\alias{ig}
\docType{package}

\title{
Package for a General Class of Inverse Gaussian Distributions
}

\description{
A collection of utilities for a general class of inverse Gaussian models known 
as inverse Gaussian type distribution (IGTD).
}

\details{
\tabular{ll}{
Package: \tab ig\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2008-04-20\cr
License: \tab GPL-2\cr
}
}

\author{
Víctor Leiva <victor.leiva@uv.cl; victor.leiva@yahoo.com>, \cr
Hugo Hernández <hugo.hernandez.p@gmail.com> and \cr
Antonio Sanhueza <asanhueza@ufro.cl>.
}

\references{
Chhikara, R. S. and Folks, J. L. (1989). The Inverse Gaussian Distribution. 
Marcel Dekker, New York.
  
Sanhueza, A., Leiva, V., Balakrishnan, N. (2008). A new class of 
inverse Gaussian type distributions. Metrika (in press).
}

\keyword{package}
