% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ig_api_get}
\alias{ig_api_get}
\title{GET Instagram API}
\usage{
ig_api_get(...)
}
\arguments{
\item{...}{Path and query string components (endpoint parameters) should be
supplied here. If unnamed, the first object is assumed to be the API path
path, which is the string pointing to the desired API endpoint
\code{users/self}. Additional named arguments supplied here will be included
as part of the query string (trailing the "?" in the URL).}
}
\value{
An HTTP response object.
}
\description{
Send GET requests to Instagram's API
}
\examples{
\dontrun{
## make custom request to locations/search endpoint
ig_locs <- ig_api_get("locations/search", lat = 48.858844, lng = 2.294351)

## view data
ig_as_tbl(ig_locs)
}
}
