% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_media.R
\name{ig_media_search}
\alias{ig_media_search}
\title{media/search}
\usage{
ig_media_search(...)
}
\arguments{
\item{...}{Passed to query string.}
}
\value{
Returns media data.
}
\description{
Interacting with the media/search endpoint of Instagram's API
}
\examples{
\dontrun{
ig_media <- ig_media_search()
ig_as_tbl(ig_media)
}
}
