% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_media.R
\name{ig_tags_media_recent}
\alias{ig_tags_media_recent}
\title{tags/{tag-name}/media/recent}
\usage{
ig_tags_media_recent(tag = "nofilter", ...)
}
\arguments{
\item{tag}{Name of tag.}

\item{...}{Passed to query string.}
}
\value{
Returns tags media data.
}
\description{
Interacting with the tags/{tag-name}/media/recent endpoint of Instagram's API
}
\examples{
\dontrun{
## get recent media
tags_media <- ig_tags_media_recent()

## view data
ig_as_tbl(tags_media)
}
}
