% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_media.R
\name{ig_tags_name}
\alias{ig_tags_name}
\title{tags/{tags-name}}
\usage{
ig_tags_name(tag = "nofilter", ...)
}
\arguments{
\item{tag}{Name of tag.}

\item{...}{Passed to query string.}
}
\value{
Returns tags data.
}
\description{
Interacting with the tags/{tags-name} endpoint of Instagram's API
}
\examples{
\dontrun{
ig_tags <- ig_tags_name()
as_tbl(ig_tags)
}
}
