% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_self.R
\name{ig_tags_search}
\alias{ig_tags_search}
\title{tags/search}
\usage{
ig_tags_search(...)
}
\value{
Returns tags data.
}
\description{
Interacting with the tags/search endpoint of Instagram's API
}
\examples{
\dontrun{
ig_tags <- ig_tags_search()
as_tbl(ig_tags)
}
}
