\name{kappaII}
\alias{kappaII}

\title{Transformation of the IGTD distribution}

\description{
Compute the value of a tranformation of the IGT distribution.
}

\usage{
kappaII(t = 1, theta = c(1, 1))
}

\arguments{
  \item{t}{Vector of observations.}
  \item{theta}{Vector of parameters of mu and lambda.}
}

\value{
\code{kappa()} compute the value of a tranformation of the IGT distribution.
}

\references{
Sanhueza, A., Leiva, V. and Balakrishnan, N. (2007). A new class of inverse Gaussian type distributions. Metrika (in press).
}

\author{
Víctor Leiva <victor.leiva@uv.cl>, Hugo Hernández <hugo.hernandez@msn.com>, and Antonio Sanhueza <asanhue@ufro.cl>.
}

\keyword{univar}
\keyword{misc}
