\name{mleIG}
\alias{mleIG}

\title{Maximum Likelihood Estimators (MLE) of the parameters of the inverse Gaussian distribution}

\description{Compute the MLE of the parameters of the inverse Gaussian distribution.}

\usage{
mleIG(x)
}

\arguments{
  \item{x}{Vector of observations.}
}

\references{
Seshadri V. (1999). The inverse Gaussian distribution: statistical theory and applications. Springer. New York.
}

\author{
Víctor Leiva <victor.leiva@uv.cl>, Hugo Hernández <hugo.hernandez@msn.com>, and Antonio Sanhueza <asanhue@ufro.cl>.
}

\keyword{math}
\keyword{htest}
