\name{psi21}

\alias{psi21}

\docType{data}

\title{Data set}

\description{
  Data set related to the IGTD which is available in the \pkg{ig} package. This data set 
  has been taken from the literature of this topic.
}

\usage{data(psi21)}

\format{
A vector containing 101 observations.
}

\details{
\code{psi21}, \code{psi26} and \code{psi31} were taken from Birnbaum and Saunders (1969), who 
reported fatigue life data correspond to the cycles (\eqn{\times 10^{-3}}) 
of aluminum specimens of type 6061-T6. These specimens were cut in a parallel 
angle to the direction of rotation and oscillating at 18 cycles per seconds. 
They were exposed to a pressure with maximum stress of 21,000, 26,000 and 
31,000 psi (pounds per square inch), for \eqn{n = 101, 102,} and \eqn{101} 
specimens, respectively.

Other data set names are: \code{repairtimes}, \code{shelflife}, \code{fracture}, \code{precipitations} 
and \code{runoff}, which correspond to maintenance data 
on active repair times (in hours) for an airborne communications transceiver, 
shelflife (in days) of a food product, fracture toughnesses of welds, 
precipitation (in inches) and runoff amounts at Jug Bridge, Maryland, 
respectively.
}

\source{
Birnbaum Z. W. and Saunders S. C. (1969). Estimation for a family
of life distributions. Journal of Applied Probability, 6, 319-327.
}

\author{
Víctor Leiva <victor.leiva@uv.cl; victor.leiva@yahoo.com>, \cr
Hugo Hernández <hugo.hernandez.p@gmail.com> and \cr
Antonio Sanhueza <asanhueza@ufro.cl>.
}

\keyword{datasets}



