\name{search.mode}
\alias{search.mode}

\title{Function for obtaining the empirical mode of the data}

\description{
Function for obtaining the empirical mode of the data.
}

\usage{
search.mode(x)
}

\arguments{
  \item{x}{Vector of observations.}
}

\author{
Víctor Leiva <victor.leiva@uv.cl>, Hugo Hernández <hugo.hernandez@msn.com>, and Antonio Sanhueza <asanhue@ufro.cl>.
}

\keyword{univar}
