\name{searchMode}

\alias{searchMode}

\title{Function for obtaining the empirical mode of the data}

\description{
  This function computes the empirical mode of the data.
}

\usage{
searchMode(x)
}

\arguments{
  \item{x}{Vector of observations.}
}

\value{
  \code{searcMode()} return the empirical mode (modal value) of the data.
}

\author{
Víctor Leiva <victor.leiva@uv.cl; victor.leiva@yahoo.com>, \cr
Hugo Hernández <hugo.hernandez.p@gmail.com> and \cr
Antonio Sanhueza <asanhueza@ufro.cl>.
}

\keyword{univar}
\keyword{math}
