\name{shelflife}
\alias{shelflife}
\docType{data}

\title{Data set}

\description{
Several data sets related to the IGTD are available in the ig package, which have been taken from the literature of this topic.}

\usage{data(shelflife)}

\format{
A vector containing 26 observations.
}

\details{
psi21, psi26 and psi31 were taken from Birnbaum and Saunders (1969), who reported fatigue life data correspond
to the cycles (\eqn{\times 10^{-3}}) of aluminum specimens of type 6061-T6.
These specimens were cut in a parallel angle to the direction of rotation and oscillating
at 18 cycles per seconds. They were exposed to a pressure with maximum stress of 21,000, 26,000,
and 31,000 psi (pounds per square inch), for \eqn{n = 101, 102,} and \eqn{101} specimens, respectively.

Other data set names are: repairtimes, shelflife, fracture, precipitations and runoff Chhikara and Folks (1989),
which correspond to maintenance data on active repair times (in hours) for an airborne
communications transceiver, shelflife (in days) of a food product, fracture toughnesses
of MIG welds and precipitation (in inches) and runoff amounts at Jug Bridge, Maryland, respectively.
}

\references{
Chhikara, R. S. and Folks, J. L. (1989). The Inverse Gaussian Distribution. Marcel Dekker, New York.
}

\keyword{datasets}
