\name{ig-package}
\alias{ig-package}
\alias{ig}
\docType{package}

\title{
Package for the robust and classical versions of the inverse Gaussian distribution
}

\description{
A collection of utilities for robust and classical versions of the inverse Gaussian distribution known as the inverse Gaussian type distribution (IGTD).
}
\details{
\tabular{ll}{
Package: \tab ig\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-09-05\cr
License: \tab GPL\cr
}
}

\author{
Víctor Leiva <victor.leiva@uv.cl; victor.leiva@yahoo.com>, Hugo Hernández <hugo.hernandez@msn.com>, and Antonio Sanhueza <asanhue@ufro.cl>

Maintainer: Hugo Hernández <hugo.hernandez@msn.com>
}

\references{
Chhikara, R. S. and Folks, J. L. (1989). The Inverse Gaussian Distribution. Marcel Dekker, New York.
}

\keyword{package}
