\name{ig-package}
\alias{ig-package}
\alias{ig}
\docType{package}

\title{
Package for the analysis of inverse Gaussian type distributions
}

\description{
A collection of utilities for inverse Gaussian type distributions (IGTD).
}

\details{
\tabular{ll}{
Package: \tab ig\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2008-04-11\cr
License: \tab GPL\cr
}
}

\author{
Víctor Leiva <victor.leiva@uv.cl; victor.leiva@yahoo.com>, \cr
Hugo Hernández <hugo.hernandez.p@gmail.com> and \cr
Antonio Sanhueza <asanhueza@ufro.cl>.
}

\references{
Chhikara, R. S. and Folks, J. L. (1989). The Inverse Gaussian Distribution. 
Marcel Dekker, New York.
  
Sanhueza, A., Leiva, V., Balakrishnan, N. (2008). A new class of 
inverse Gaussian type distributions. Metrika (in press).
}

\keyword{package}
