% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_media.R
\name{ig_tags_search}
\alias{ig_tags_search}
\title{tags/search}
\usage{
ig_tags_search(q, ...)
}
\arguments{
\item{q}{Search query.}

\item{...}{Passed to query string.}
}
\value{
Returns tags data.
}
\description{
Interacting with the tags/search endpoint of Instagram's API
}
\examples{
\dontrun{
ig_tags <- ig_tags_search()
ig_as_tbl(ig_tags)
}
}
