% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_self.R
\name{ig_users_self}
\alias{ig_users_self}
\title{users/self}
\usage{
ig_users_self()
}
\value{
Returns user data.
}
\description{
Interacting with the users/self endpoint of Instagram's API
}
\examples{
\dontrun{
## get your data
my_ig <- ig_users_self()

## view data
ig_as_tbl(my_ig)
}
}
