\name{kappaii}

\alias{kappaii}

\title{Transformation of the IGTD distribution}

\description{
  This function gives the value of a tranformation of the IGT distribution.
}

\usage{
kappaii(x, theta = c(1.0, 1.0))
}

\arguments{
  \item{x}{Vector of observations.}
  \item{theta}{Vector of parameters of mu and lambda.}
}

\value{
  \code{kappaii()} computes the value of a tranformation of the IGTD.
}

\references{Sanhueza, A., Leiva, V., Balakrishnan, N. (2008). A new class of 
            inverse Gaussian type distributions. Metrika (in press).
}

\author{
Víctor Leiva <victor.leiva@uv.cl; victor.leiva@yahoo.com>, \cr
Hugo Hernández <hugo.hernandez.p@gmail.com> and \cr
Antonio Sanhueza <asanhueza@ufro.cl>.
}

\keyword{univar}
\keyword{misc}
