\name{mleIGt}
\alias{mleIGt}

\title{Maximum Likelihood Estimators (MLE) of the parameters of the inverse Gaussian type distribution generated from the t kernel}

\description{Compute the MLE of the parameters of the inverse Gaussian type distribution generated from the t kernel.}

\usage{
mleIGt(x, nu.fixed = 2)
}

\arguments{
  \item{x}{Vector of observations.}
  \item{nu.fixed}{Parameter of the IGTD when the t kernel is used.}
}

\references{
Sanhueza, A., Leiva, V. and Balakrishnan, N. (2007). A new class of inverse Gaussian type distributions. Metrika (in press).
}

\author{
Víctor Leiva <victor.leiva@uv.cl>, Hugo Hernández <hugo.hernandez@msn.com>, and Antonio Sanhueza <asanhue@ufro.cl>.
}

\keyword{math}
\keyword{htest}
